8/23 최신화(김희진 - 김현성)

+++ src/main/java/com/Daldagu1/TeamDaldagu1Project/controller/RestApiController.java
@@ -28,17 +28,18 @@ public class RestApiController {
 
     @PostMapping("/rest/approve_seller")
     public void approveSeller(@RequestParam("user_idx") int user_idx){
-        SellerInfoBean tempSellerInfoBean = sellerService.getSellerInfo(user_idx);
+        if(userService.getUserbyIdx(user_idx).getUser_role().equals("S")){}
+        else {
+            SellerInfoBean tempSellerInfoBean = sellerService.getSellerInfo(user_idx);
 
-        SellerBean tempSellerBean = new SellerBean();
-        tempSellerBean.setSeller_idx(tempSellerInfoBean.getSeller_idx());
-        tempSellerBean.setCalc_account(tempSellerInfoBean.getCalc_account());
-        tempSellerBean.setSelling_page_title(tempSellerInfoBean.getSeller_page_title());
-        tempSellerBean.setUser_idx(user_idx);
+            SellerBean tempSellerBean = new SellerBean();
+            tempSellerBean.setCalc_account(tempSellerInfoBean.getCalc_account());
+            tempSellerBean.setSelling_page_title(tempSellerInfoBean.getSeller_page_title());
+            tempSellerBean.setUser_idx(user_idx);
 
-        sellerService.addSeller(tempSellerBean);
-        sellerService.deleteSellerJoinInfo(tempSellerBean.getUser_idx());
-
+            sellerService.addSeller(tempSellerBean);
+            sellerService.deleteSellerJoinInfo(tempSellerBean.getUser_idx());
+        }
     }
 
     @PostMapping("/rest/denial_seller")
